Separated gcrhenv on GCToEEInterface.cpp and RedhawkGCInterface.cpp
